requirements.txt
customtkinter
pandas
openpyxl
Pillow
tkinterdnd2-universal
chardet 